+++ sample/Pizzeria.Store.Web/Pizzeria.Store.Web.csproj
@@ -16,6 +16,7 @@
     <ProjectReference Include="..\..\src\Lewee.Blazor\Lewee.Blazor.csproj" />
     <ProjectReference Include="..\..\src\Lewee.Infrastructure.AspNet\Lewee.Infrastructure.AspNet.csproj" />
     <ProjectReference Include="..\..\src\Lewee.Infrastructure.Auth\Lewee.Infrastructure.Auth.csproj" />
+    <ProjectReference Include="..\..\src\Lewee.Infrastructure.Correlate\Lewee.Infrastructure.Correlate.csproj" />
     <ProjectReference Include="..\..\src\Lewee.Infrastructure.PostgreSQL\Lewee.Infrastructure.PostgreSQL.csproj" />
     <ProjectReference Include="..\Pizzeria.Common\Pizzeria.Common.csproj" />
     <ProjectReference Include="..\Pizzeria.Store.Application\Pizzeria.Store.Application.csproj" />
