+++ src/Lewee.Infrastructure.Correlate/Lewee.Infrastructure.Correlate.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <PackageReference Include="Correlate.AspNetCore" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Lewee.Common\Lewee.Common.csproj" />
+  </ItemGroup>
+
+  <!-- No functionality, just helping setup `Correlate` -->
+  <ItemGroup>
+    <AssemblyAttribute Include="System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverageAttribute" />
+  </ItemGroup>
+
+</Project>
